refactor: resolve module structure conflicts and unify dependencies
- Resolved systemic architectural debt where flat `.rs` files coexisted with same-named directories, causing compiler ambiguity.
- Consolidated redundant modules: `episodic` merged into `episode`, and `pattern` merged into `patterns` in `do-memory-core`.
- Moved `embeddings_simple.rs` to `embeddings/simple.rs`.
- Unified `sysinfo` dependency to version 0.38 across the workspace to satisfy Rust 1.94.0 MSRV constraints.
- Fixed `#[path]` attribute regressions in `memory-storage-turso` after moving files.
- Updated all internal crate references and imports to reflect the new module structure.
- Resolved visibility issues and module declaration conflicts discovered during build verification.
- Synced `Cargo.lock` with unified dependency versions.

Structural changes verified via file system audit and `cargo update`. Full build verification is blocked by a known environment-specific Clang segmentation fault in the `aegis` dependency.
