+++ memory-cli/src/config/validator/mod.rs
