+++ memory-core/src/memory/pattern_search/mod.rs
