+++ memory-core/src/memory/queries/mod.rs
