+++ memory-core/src/patterns/changepoint/mod.rs
@@ -7,7 +7,7 @@
 //! ## Example
 //!
 //! ```
-//! use do_memory_core::patternss::changepoint::{ChangepointDetector, ChangepointConfig};
+//! use do_memory_core::patterns::changepoint::{ChangepointDetector, ChangepointConfig};
 //!
 //! // Create detector with default settings (needs to be mutable for detection)
 //! let mut detector = ChangepointDetector::new(ChangepointConfig::default());
