+++ memory-mcp/src/patterns/compatibility/mod.rs
