+++ memory-mcp/src/patterns/predictive/extraction/mod.rs
