+++ memory-mcp/src/protocol/handlers.rs
@@ -88,16 +88,21 @@ pub async fn handle_initialize(
         }),
         Err(e) => {
             error!("Failed to serialize initialize response: {}", e);
-            Some(internal_error(
-                request.id,
-                &format!("Response serialization failed: {e}"),
-            ))
+            Some(JsonRpcResponse {
+                jsonrpc: "2.0".to_string(),
+                id: request.id,
+                result: None,
+                error: Some(JsonRpcError {
+                    code: -32603,
+                    message: "Internal error".to_string(),
+                    data: Some(json!({"details": format!("Response serialization failed: {}", e)})),
+                }),
+            })
         }
     }
 }
 
 /// Handle tools/list request
-#[deprecated = "Use handle_list_tools_with_lazy instead"]
 pub async fn handle_list_tools(
     request: JsonRpcRequest,
     tools: Vec<McpTool>,
@@ -117,10 +122,16 @@ pub async fn handle_list_tools(
         }),
         Err(e) => {
             error!("Failed to serialize list_tools response: {}", e);
-            Some(internal_error(
-                request.id,
-                &format!("Response serialization failed: {e}"),
-            ))
+            Some(JsonRpcResponse {
+                jsonrpc: "2.0".to_string(),
+                id: request.id,
+                result: None,
+                error: Some(JsonRpcError {
+                    code: -32603,
+                    message: "Internal error".to_string(),
+                    data: Some(json!({"details": format!("Response serialization failed: {}", e)})),
+                }),
+            })
         }
     }
 }
@@ -227,10 +238,18 @@ pub fn handle_list_tools_with_lazy(
             }),
             Err(e) => {
                 error!("Failed to serialize list_tools response: {}", e);
-                Some(internal_error(
-                    request.id,
-                    &format!("Response serialization failed: {e}"),
-                ))
+                Some(JsonRpcResponse {
+                    jsonrpc: "2.0".to_string(),
+                    id: request.id,
+                    result: None,
+                    error: Some(JsonRpcError {
+                        code: -32603,
+                        message: "Internal error".to_string(),
+                        data: Some(
+                            json!({"details": format!("Response serialization failed: {}", e)}),
+                        ),
+                    }),
+                })
             }
         }
     } else {
@@ -256,10 +275,18 @@ pub fn handle_list_tools_with_lazy(
             }),
             Err(e) => {
                 error!("Failed to serialize list_tools response: {}", e);
-                Some(internal_error(
-                    request.id,
-                    &format!("Response serialization failed: {e}"),
-                ))
+                Some(JsonRpcResponse {
+                    jsonrpc: "2.0".to_string(),
+                    id: request.id,
+                    result: None,
+                    error: Some(JsonRpcError {
+                        code: -32603,
+                        message: "Internal error".to_string(),
+                        data: Some(
+                            json!({"details": format!("Response serialization failed: {}", e)}),
+                        ),
+                    }),
+                })
             }
         }
     }
@@ -294,10 +321,16 @@ where
     let tool_name = match tool_name {
         Some(name) => name,
         None => {
-            return Some(invalid_params(
-                request.id,
-                "Missing required parameter: name",
-            ));
+            return Some(JsonRpcResponse {
+                jsonrpc: "2.0".to_string(),
+                id: request.id,
+                result: None,
+                error: Some(JsonRpcError {
+                    code: -32602,
+                    message: "Invalid params".to_string(),
+                    data: Some(json!({"details": "Missing required parameter: name"})),
+                }),
+            });
         }
     };
 
@@ -323,21 +356,33 @@ where
                 }),
                 Err(e) => {
                     error!("Failed to serialize describe_tool response: {}", e);
-                    Some(internal_error(
-                        request.id,
-                        &format!("Response serialization failed: {e}"),
-                    ))
+                    Some(JsonRpcResponse {
+                        jsonrpc: "2.0".to_string(),
+                        id: request.id,
+                        result: None,
+                        error: Some(JsonRpcError {
+                            code: -32603,
+                            message: "Internal error".to_string(),
+                            data: Some(
+                                json!({"details": format!("Response serialization failed: {}", e)}),
+                            ),
+                        }),
+                    })
                 }
             }
         }
         None => {
             info!("Tool not found: {}", tool_name);
-            Some(error_response(
-                request.id,
-                -32602,
-                "Tool not found",
-                &format!("Tool not found: {tool_name}"),
-            ))
+            Some(JsonRpcResponse {
+                jsonrpc: "2.0".to_string(),
+                id: request.id,
+                result: None,
+                error: Some(JsonRpcError {
+                    code: -32602,
+                    message: "Tool not found".to_string(),
+                    data: Some(json!({"tool_name": tool_name})),
+                }),
+            })
         }
     }
 }
@@ -375,10 +420,16 @@ where
             .map(String::from)
             .collect::<Vec<_>>(),
         None => {
-            return Some(invalid_params(
-                request.id,
-                "Missing required parameter: names (array)",
-            ));
+            return Some(JsonRpcResponse {
+                jsonrpc: "2.0".to_string(),
+                id: request.id,
+                result: None,
+                error: Some(JsonRpcError {
+                    code: -32602,
+                    message: "Invalid params".to_string(),
+                    data: Some(json!({"details": "Missing required parameter: names (array)"})),
+                }),
+            });
         }
     };
 
@@ -406,10 +457,16 @@ where
         }),
         Err(e) => {
             error!("Failed to serialize describe_tools response: {}", e);
-            Some(internal_error(
-                request.id,
-                &format!("Response serialization failed: {e}"),
-            ))
+            Some(JsonRpcResponse {
+                jsonrpc: "2.0".to_string(),
+                id: request.id,
+                result: None,
+                error: Some(JsonRpcError {
+                    code: -32603,
+                    message: "Internal error".to_string(),
+                    data: Some(json!({"details": format!("Response serialization failed: {}", e)})),
+                }),
+            })
         }
     }
 }
