+++ memory-storage-redb/src/episodes_summaries.rs
@@ -1,7 +1,7 @@
 //! Episode summaries and capacity operations for redb cache
 
 use crate::{EPISODES_TABLE, METADATA_TABLE, RedbStorage, SUMMARIES_TABLE};
-use do_memory_core::episodic::CapacityManager;
+use do_memory_core::episode::CapacityManager;
 use do_memory_core::semantic::EpisodeSummary;
 use do_memory_core::{Episode, Error, Result};
 use redb::{ReadableDatabase, ReadableTable, ReadableTableMetadata};
@@ -230,7 +230,7 @@ impl RedbStorage {
     /// ```no_run
     /// # use do_memory_storage_redb::RedbStorage;
     /// # use do_memory_core::{Episode, TaskContext, TaskType};
-    /// # use do_memory_core::episodic::{CapacityManager, EvictionPolicy};
+    /// # use do_memory_core::episode::{CapacityManager, EvictionPolicy};
     /// # use std::path::Path;
     /// # async fn example() -> anyhow::Result<()> {
     /// # let storage = RedbStorage::new(Path::new("./test.redb")).await?;
